dtype fixes

+++ actor_critic/actor.py
@@ -36,7 +36,7 @@ class Actor(ActorCriticBase):
     def forward(self, observation: torch.Tensor) -> torch.Tensor:
         assert observation.ndim >= 2
         assert observation.shape[-2:] == (self._history_size, self._input_features)
-        result = self._forward_model_a(observation.float().flatten(-2, -1))  # TODO
+        result = self._forward_model_a(observation.flatten(-2, -1))  # TODO
         assert result.shape == observation.shape[:-2] + (self._output_features,)
         return result
